requirements-dev.txt
python-dotenv # REMOVE ON DEPLOYMENT
google-generativeai
unsloth # only works on linux
ollama